+++ client/src/utils/mask-credential.ts
@@ -1,4 +1,0 @@
-export const maskCredential = (credential: string | null): string => {
-  const displayCredential = `(${credential?.substring(0, 3)}...${credential?.substring(credential?.length - 3)})`
-  return displayCredential
-}